Refined Store implementation
